change name of black flag harness for readability
