+++ games/ac_blackflag_resynced/ac_blackflag_resynced.py
@@ -1,11 +1,10 @@
 """Assassins Creed Black Flag Resynced test script"""
 
 import logging
+import re
 import sys
 import time
 from pathlib import Path
-
-from acbfr_utils import get_resolution
 
 PARENT_DIRECTORY = str(Path(__file__).resolve().parent.parent.parent)
 sys.path.insert(1, PARENT_DIRECTORY)
@@ -51,6 +50,34 @@ intro_videos = [
     VIDEO_PATH / "en" / "warning_disclaimer.webm",
     VIDEO_PATH / "en" / "WarningSaving.webm",
 ]
+
+
+def get_resolution():
+    """Get current resolution from settings file"""
+    config_path = (
+        user_documents(STEAM_GAME_ID)
+        / "Assassin's Creed Black Flag Resynced"
+        / "ACBlackFlag.ini"
+    )
+    if not config_path.exists():
+        raise RuntimeError(f"Missing path: {config_path}")
+    logger.info("Reading AC Black Flag Resynced settings file path=%s", config_path)
+    height_pattern = re.compile(r"FullscreenHeight=(\d+)")
+    width_pattern = re.compile(r"FullscreenWidth=(\d+)")
+    height = None
+    width = None
+    with open(config_path, encoding="utf-8") as f:
+        lines = f.readlines()
+        for line in lines:
+            height_match = height_pattern.search(line)
+            width_match = width_pattern.search(line)
+            if height_match is not None:
+                height = int(height_match.group(1))
+            if width_match is not None:
+                width = int(width_match.group(1))
+            if height is not None and width is not None:
+                return (width, height)
+    return (width, height)
 
 
 def navigate_to_settings():
@@ -137,7 +164,7 @@ def navigate_settings() -> None:
     capture_and_save_screenshot(ARTIFACTS_DIRECTORY / "graphics_3.png")
 
     press("down*5")
-    
+
     result = find_word("texture", interval=3, timeout=20)
     if not result:
         logger.info(
